requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
dj-database-url==0.5.0
Django==3.2.7
gunicorn==20.1.0
idna==3.2
numpy==1.21.2
Pillow==8.3.2
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
torch==1.9.0+cpu
torchvision==0.10.0+cpu
typing-extensions==3.10.0.2
urllib3==1.26.6
whitenoise==5.3.0
